feat: PromotionModal에 포커스 이동 및 접근성 개선
