+++ src/hooks/FocusTrap.tsx
@@ -1,44 +1,8 @@
-import { useCallback, useLayoutEffect, useRef } from 'react';
+import { useLayoutEffect, useRef } from 'react';
+import { getFocusableElements } from '../utils/getFocusableElements';
 
 function FocusTrap({ children }: { children: React.ReactNode }) {
   const modalRef = useRef<HTMLDivElement | null>(null);
-
-  // https://allyjs.io/data-tables/focusable.html
-  // https://html.spec.whatwg.org/multipage/interaction.html#focusable-area
-  // HTML 스펙 기준으로 포커스 가능한 요소를 찾는 함수
-
-  const getFocusableElements = useCallback((element: HTMLElement) => {
-    const focusableSelectors = [
-      'a[href]',
-      'area[href]',
-      'input:not([disabled])',
-      'select:not([disabled])',
-      'textarea:not([disabled])',
-      'button:not([disabled])',
-      'iframe',
-      'object',
-      'embed',
-      '[contenteditable]',
-      '[tabindex]:not([tabindex="-1"])'
-    ];
-
-    const focusableElements = Array.from(
-      element.querySelectorAll<HTMLElement>(focusableSelectors.join(','))
-    ).filter((el) => {
-      // aria-hidden이나 hidden 속성이 있는 요소 제외
-      return (
-        !el.getAttribute('aria-hidden') &&
-        !el.getAttribute('aria-disabled') &&
-        !el.getAttribute('disabled') &&
-        !el.getAttribute('readonly') &&
-        !el.hasAttribute('hidden') &&
-        el.style.display !== 'none' &&
-        el.style.visibility !== 'hidden'
-      );
-    });
-
-    return focusableElements;
-  }, []);
 
   useLayoutEffect(() => {
     const element = modalRef.current;
@@ -75,7 +39,7 @@ function FocusTrap({ children }: { children: React.ReactNode }) {
     return () => {
       document.removeEventListener('keydown', handleKeyDown);
     };
-  }, [getFocusableElements]);
+  }, []);
 
   return <div ref={modalRef}>{children}</div>;
 }
